[MOD] moved help page popup content inside their own components
